Check in results of wayland-scanner
It is customary for Go packages to include generated code, so that "go
build" just works, not requiring to run custom makefiles. This is
especially important for packages that are meant to be used as a
dependency (like this one), as otherwise building the dependent will
not work without modifying its dependency.