Add Go version to go.mod
go build automatically adds it, so check in the result.

We could enforce a lower version, but I don't see much use in that. We
can't enforce Go 1.8 because of a bug in cgo – running a newer Go will
always use type aliases in cgo output, but 1.8 didn't support type
aliases. And we should probably increase the minimum required version
at some point – Go 1.8 is from the stone ages.