Provide typesafe abstraction over input device union
Don't expose the user to unions, which circumvent memory safety.
Instead, introduce an InputDevice interface, and make sure that all
devices are of a concrete type providing access to only the correct
field of the union.

In theory, we could drop the Type method, as users can now rely on
type switches. However, we don't currently have unique types for every
kind of device, and we may not keep up with new types being added to
wlroots, so the Type method provides a useful fallback, at least for
debugging.